refactor: split app_theme into app_theme, app_colors, and app_text_styles
