+++ lib/src/app.dart
@@ -8,6 +8,7 @@ import 'package:nokken/src/core/theme/providers/theme_provider.dart';
 import 'package:nokken/src/core/services/navigation/routes/app_router.dart';
 import 'package:nokken/src/core/screens/main_screen.dart';
 import 'package:nokken/src/core/theme/app_theme.dart';
+import 'package:nokken/src/core/theme/app_colors.dart';
 
 /// Root widget for the app
 /// Handles theme configuration and routing
