+++ lib/src/core/screens/main_screen.dart
@@ -9,8 +9,9 @@ import 'package:nokken/src/features/settings/screens/settings_screen.dart';
 import 'package:nokken/src/core/theme/shared_widgets.dart';
 import 'package:nokken/src/features/scheduler/screens/daily_tracker_screen.dart';
 import 'package:nokken/src/features/medication_tracker/screens/medication_list_screen.dart';
-import 'package:nokken/src/core/theme/app_icons.dart';
 import 'package:nokken/src/core/theme/app_theme.dart';
+import 'package:nokken/src/core/theme/app_colors.dart';
+import 'package:nokken/src/core/theme/app_icons.dart';
 
 /// Provider to track the current navigation index
 /// Default to index 2 (daily tracker)
