+++ lib/src/core/theme/app_theme.dart
@@ -4,197 +4,7 @@
 //
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-
-//----------------------------------------------------------------------------
-// APP COLORS
-//----------------------------------------------------------------------------
-
-/// Theme-aware color provider that adjusts based on current theme mode
-class AppColors {
-  // Private constructor to prevent instantiation
-  AppColors._();
-
-  // Current theme mode
-  static ThemeMode _themeMode = ThemeMode.dark;
-
-  // Dark and light color schemes
-  static final ColorScheme _darkScheme = AppTheme.darkColorScheme;
-  static final ColorScheme _lightScheme = AppTheme.lightColorScheme;
-
-  /// Get current scheme based on theme mode
-  static ColorScheme get current =>
-      _themeMode == ThemeMode.light ? _lightScheme : _darkScheme;
-
-  /// Method to change theme mode
-  static void setThemeMode(ThemeMode mode) {
-    _themeMode = mode;
-    // A notifier could be added here for more reactive theme changes
-  }
-
-  // Basic theme colors
-  static Color get primary => current.primary;
-  static Color get onPrimary => current.onPrimary;
-  static Color get secondary => current.secondary;
-  static Color get onSecondary => current.onSecondary;
-  static Color get tertiary => current.tertiary;
-  static Color get onTertiary => current.onTertiary;
-
-  // Surface colors
-  static Color get surface => current.surface;
-  static Color get surfaceContainer => current.surfaceContainer;
-  static Color get onSurface => current.onSurface;
-  static Color get onSurfaceVariant => current.onSurfaceVariant;
-
-  // Status colors
-  static Color get error => current.error;
-  static Color get errorContainer => current.errorContainer;
-  static Color get success =>
-      _themeMode == ThemeMode.light ? AppTheme.greenLight : AppTheme.greenDark;
-  static Color get warning => _themeMode == ThemeMode.light
-      ? AppTheme.orangeLight
-      : AppTheme.orangeDark;
-  static Color get info => current.primary;
-
-  // Medication type colors
-  static Color get oralMedication => _themeMode == ThemeMode.light
-      ? AppTheme.oralMedColorLight
-      : AppTheme.oralMedColorDark;
-
-  static Color get topical => _themeMode == ThemeMode.light
-      ? AppTheme.topicalColorLight
-      : AppTheme.topicalColorDark;
-
-  static Color get patch => _themeMode == ThemeMode.light
-      ? AppTheme.patchColorLight
-      : AppTheme.patchColorDark;
-
-  static Color get injection => _themeMode == ThemeMode.light
-      ? AppTheme.injectionColorLight
-      : AppTheme.injectionColorDark;
-
-  // Appointment type colors
-  static Color get bloodwork => _themeMode == ThemeMode.light
-      ? AppTheme.bloodworkColorLight
-      : AppTheme.bloodworkColorDark;
-
-  static Color get doctorAppointment => _themeMode == ThemeMode.light
-      ? AppTheme.doctorApptColorLight
-      : AppTheme.doctorApptColorDark;
-
-  static Color get surgery => _themeMode == ThemeMode.light
-      ? AppTheme.surgeryColorLight
-      : AppTheme.surgeryColorDark;
-
-  // Other UI element colors
-  static Color get shadow => current.shadow;
-  static Color get outline => current.outline;
-  static Color get cardColor => current.surfaceContainer;
-}
-
-//----------------------------------------------------------------------------
-// TEXT STYLES
-//----------------------------------------------------------------------------
-
-/// Theme-aware text styles that adjust based on current theme mode
-class AppTextStyles {
-  // Private constructor to prevent instantiation
-  AppTextStyles._();
-
-  /// Get current text theme based on theme mode
-  static TextTheme get _current => AppColors._themeMode == ThemeMode.light
-      ? _getLightTextTheme()
-      : _getDarkTextTheme();
-
-  /// Create the dark theme text styles
-  static TextTheme _getDarkTextTheme() => TextTheme(
-        displayLarge: AppTheme.displayLarge,
-        displayMedium: AppTheme.displayMedium,
-        displaySmall: AppTheme.displaySmall,
-        headlineLarge: AppTheme.headlineLarge,
-        headlineMedium: AppTheme.headlineMedium,
-        headlineSmall: AppTheme.headlineSmall,
-        titleLarge: AppTheme.titleLarge,
-        titleMedium: AppTheme.titleMedium,
-        titleSmall: AppTheme.titleSmall,
-        bodyLarge: AppTheme.bodyLarge,
-        bodyMedium: AppTheme.bodyMedium,
-        bodySmall: AppTheme.bodySmall,
-        labelLarge: AppTheme.labelLarge,
-        labelMedium: AppTheme.labelMedium,
-        labelSmall: AppTheme.labelSmall,
-      );
-
-  /// Create the light theme text styles with adjusted colors
-  static TextTheme _getLightTextTheme() {
-    return TextTheme(
-      displayLarge: AppTheme.displayLarge.copyWith(color: AppTheme.black),
-      displayMedium: AppTheme.displayMedium.copyWith(color: AppTheme.black),
-      displaySmall: AppTheme.displaySmall.copyWith(color: AppTheme.black),
-      headlineLarge: AppTheme.headlineLarge.copyWith(color: AppTheme.black),
-      headlineMedium: AppTheme.headlineMedium.copyWith(color: AppTheme.black),
-      headlineSmall: AppTheme.headlineSmall.copyWith(color: AppTheme.black),
-      titleLarge: AppTheme.titleLarge.copyWith(color: AppTheme.black),
-      titleMedium: AppTheme.titleMedium.copyWith(color: AppTheme.black),
-      titleSmall: AppTheme.titleSmall.copyWith(color: AppTheme.black),
-      bodyLarge: AppTheme.bodyLarge.copyWith(color: AppTheme.black),
-      bodyMedium: AppTheme.bodyMedium.copyWith(color: AppTheme.black),
-      bodySmall: AppTheme.bodySmall.copyWith(color: AppTheme.black),
-      labelLarge: AppTheme.labelLarge.copyWith(color: AppTheme.black),
-      labelMedium: AppTheme.labelMedium.copyWith(color: AppTheme.black),
-      labelSmall: AppTheme.labelSmall.copyWith(color: AppTheme.black),
-    );
-  }
-
-  // Display styles
-  static TextStyle get displayLarge => _current.displayLarge!;
-  static TextStyle get displayMedium => _current.displayMedium!;
-  static TextStyle get displaySmall => _current.displaySmall!;
-
-  // Headline styles
-  static TextStyle get headlineLarge => _current.headlineLarge!;
-  static TextStyle get headlineMedium => _current.headlineMedium!;
-  static TextStyle get headlineSmall => _current.headlineSmall!;
-
-  // Title styles
-  static TextStyle get titleLarge => _current.titleLarge!;
-  static TextStyle get titleMedium => _current.titleMedium!;
-  static TextStyle get titleSmall => _current.titleSmall!;
-
-  // Body styles
-  static TextStyle get bodyLarge => _current.bodyLarge!;
-  static TextStyle get bodyMedium => _current.bodyMedium!;
-  static TextStyle get bodySmall => _current.bodySmall!;
-
-  // Label styles
-  static TextStyle get labelLarge => _current.labelLarge!;
-  static TextStyle get labelMedium => _current.labelMedium!;
-  static TextStyle get labelSmall => _current.labelSmall!;
-
-  // Utility styles - these adjust color based on current theme automatically
-  static TextStyle get buttonText => AppColors._themeMode == ThemeMode.light
-      ? AppTheme.buttonText.copyWith(color: AppTheme.black)
-      : AppTheme.buttonText;
-
-  static TextStyle get caption => AppColors._themeMode == ThemeMode.light
-      ? AppTheme.caption.copyWith(color: AppTheme.black.withAlpha(179))
-      : AppTheme.caption;
-
-  static TextStyle get overline => AppColors._themeMode == ThemeMode.light
-      ? AppTheme.overline.copyWith(color: AppTheme.black)
-      : AppTheme.overline;
-
-  static TextStyle get error => AppColors._themeMode == ThemeMode.light
-      ? AppTheme.error.copyWith(color: AppTheme.pinkDark.withRed(220))
-      : AppTheme.error;
-
-  static TextStyle get link => AppColors._themeMode == ThemeMode.light
-      ? AppTheme.link.copyWith(color: AppTheme.blueLight)
-      : AppTheme.link;
-}
-
-//----------------------------------------------------------------------------
-// MAIN THEME CONFIGURATION
-//----------------------------------------------------------------------------
+import 'package:nokken/src/core/theme/app_colors.dart';
 
 /// Main theme class with comprehensive style definitions
 class AppTheme {
