+++ lib/src/features/medication_tracker/screens/add_edit_medication_screen.dart
@@ -12,6 +12,8 @@ import 'package:nokken/src/core/constants/date_constants.dart';
 import 'package:nokken/src/core/utils/date_time_formatter.dart';
 import 'package:nokken/src/core/theme/app_icons.dart';
 import 'package:nokken/src/core/theme/app_theme.dart';
+import 'package:nokken/src/core/theme/app_colors.dart';
+import 'package:nokken/src/core/theme/app_text_styles.dart';
 import 'package:nokken/src/core/theme/shared_widgets.dart';
 import '../models/medication.dart';
 import '../providers/medication_state.dart';
