+++ lib/src/features/scheduler/screens/calendar_screen.dart
@@ -18,6 +18,8 @@ import 'package:nokken/src/core/services/database/database_service.dart';
 import 'package:nokken/src/core/services/navigation/navigation_service.dart';
 import 'package:nokken/src/core/theme/app_icons.dart';
 import 'package:nokken/src/core/theme/app_theme.dart';
+import 'package:nokken/src/core/theme/app_colors.dart';
+import 'package:nokken/src/core/theme/app_text_styles.dart';
 import 'package:nokken/src/core/utils/get_labels.dart';
 
 class CalendarScreen extends ConsumerStatefulWidget {
