+++ lib/src/features/scheduler/screens/daily_tracker_screen.dart
@@ -14,6 +14,8 @@ import 'package:nokken/src/core/services/navigation/navigation_service.dart';
 import 'package:nokken/src/core/theme/shared_widgets.dart';
 import 'package:nokken/src/core/theme/app_theme.dart';
 import 'package:nokken/src/core/theme/app_icons.dart';
+import 'package:nokken/src/core/theme/app_colors.dart';
+import 'package:nokken/src/core/theme/app_text_styles.dart';
 import 'package:nokken/src/core/utils/date_time_formatter.dart';
 import 'package:nokken/src/core/utils/get_labels.dart';
 import 'package:nokken/src/core/utils/get_icons_colors.dart';
